Updated useFollowupActiveLeads query to join on sessions.id using bigint instead of int to accommodate large WhatsApp session IDs and prevent out-of-range errors.

X-Lovable-Edit-ID: edt-c7821a68-86f9-493c-86b3-dc28470b3a1d

+++ src/pages/crm/hooks/useFollowupActiveLeads.ts
@@ -32,7 +32,7 @@
                ORDER BY fq.created_at DESC
              ) as rn
            FROM vw_send_followup_queue fq
-           INNER JOIN sessions s ON s.id = fq.session_id::int
+          INNER JOIN sessions s ON s.id = fq.session_id::bigint
            WHERE fq.cod_agent::text = ANY($1::varchar[])
              AND fq.state = 'SEND'
          )
